Refactor projection support into its own class
